+++ src/@data/withTransactions/index.js
@@ -1,16 +1,40 @@
 import { graphql } from 'react-apollo';
 import fetchMoreResolver from '@data/utils/fetchMoreResolver';
-import transactionsQuery from './transactionsQuery';
+import gql from 'graphql-tag';
 
-export default graphql(transactionsQuery, {
+export const QUERY = gql`
+  query GetTransactions($limit: Int, $skip: Int, $people: [Int], $start: String, $end: String) {
+    transactions(
+      limit: $limit,
+      skip: $skip,
+      people: $people,
+      start: $start,
+      end: $end,
+      cache: false
+    ) {
+      id
+      date
+      status
+      summary
+      person { firstName, lastName, photo }
+      details {
+        id
+        amount
+        account { id, name }
+      }
+    }
+  }
+`;
+
+export default graphql(QUERY, {
   options: (ownProps = {}) => ({
     variables: {
       limit: ownProps.limit || 20,
       skip: ownProps.skip || 0,
     },
   }),
-  props: ({ ownProps, data } = {}) => ({
-    content: data.content,
+  props: ({ ownProps, data } = {}) => console.log(data, ownProps) || ({
+    transactions: data.transactions,
     isLoading: ownProps.isLoading || data.loading,
     fetchMore: fetchMoreResolver({
       collectionName: 'content',
